requirements.txt
mss
pillow
numpy
opencv-python
py-applescript